Rename: Refactor FarmWorkModel in calendar screen
