+++ feature/main/src/main/java/kr/co/main/mapper/calendar/FarmWorkModelMapper.kt
@@ -1,8 +1,8 @@
-package kr.co.main.calendar.mapper
+package kr.co.main.mapper.calendar
 
 import kr.co.common.mapper.Mapper
 import kr.co.domain.entity.FarmWorkEntity
-import kr.co.main.calendar.model.FarmWorkModel
+import kr.co.main.model.calendar.FarmWorkModel
 
 internal object FarmWorkModelMapper
     : Mapper<FarmWorkEntity, FarmWorkModel> {
@@ -14,7 +14,7 @@ internal object FarmWorkModelMapper
                 startEra = startEra,
                 endMonth = endMonth,
                 endEra = endEra,
-                category = category,
+                type = type,
                 farmWork = farmWork,
                 videoUrl = videoUrl
             )
